- added tab bar controller - properly uploading / viewing documents - added PDF view controller - updated user profile struct
