+++ Project Doggystyle iOS/View Controllers/Tab Bar/FourthViewController.swift
@@ -1,0 +1,17 @@
+//
+//  FourthViewController.swift
+//  Project Doggystyle iOS
+//
+//  Created by Stanley Miller on 6/2/21.
+//
+
+import UIKit
+
+class FourthViewController: UIViewController {
+    var homeController: UIViewController!
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        view.backgroundColor = .systemGreen
+    }
+}
