+++ Project Doggystyle iOS/View Controllers/Tab Bar/SecondaryViewController.swift
@@ -1,0 +1,17 @@
+//
+//  SecondaryViewController.swift
+//  Project Doggystyle iOS
+//
+//  Created by Stanley Miller on 6/2/21.
+//
+
+import UIKit
+
+class SecondaryViewController: UIViewController {
+    var homeController: UIViewController!
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        view.backgroundColor = .systemPink
+    }
+}
