+++ Project Doggystyle iOS/View Controllers/Tab Bar/ThirdViewController.swift
@@ -1,0 +1,17 @@
+//
+//  ThirdViewController.swift
+//  Project Doggystyle iOS
+//
+//  Created by Stanley Miller on 6/2/21.
+//
+
+import UIKit
+
+class ThirdViewController: UIViewController {
+    var homeController: UIViewController!
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        view.backgroundColor = .systemTeal
+    }
+}
